requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
django-debug-toolbar==4.3.0
isort==5.13.2
pillow==10.3.0
celery[redis]==5.2.7
django-redis==5.4.0
flower==1.2.0
django-simple-history==3.0.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
djoser==2.2.0
django-filter==22.1